fix: process conditions after the paginator
Reordering this allows for an optimization in some cases.

For example, when using an ES search here, you'll want to do the query and load the possible filter values in the same query. The query needs limit/offset for the items, but doesn't need it for filter values. Making the paginator run first allows us to inject the current limit/offset in the query early and we don't need to run it twice.